// SDKs/googleapis.com/civicinfo/v2/typescript/src/sdk/models/shared/geocodingsummary.ts
export class GeocodingSummary extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=addressUnderstood" })
  addressUnderstood?: boolean;

  @SpeakeasyMetadata({ data: "json, name=featureId" })
  featureId?: FeatureIdProto;

  @SpeakeasyMetadata({ data: "json, name=featureType" })
  featureType?: GeocodingSummaryFeatureTypeEnum;

  @SpeakeasyMetadata({ data: "json, name=positionPrecisionMeters" })
  positionPrecisionMeters?: number;

  @SpeakeasyMetadata({ data: "json, name=queryString" })
  queryString?: string;
}